+++ libraries/interop/dev_deps.ts
@@ -2,3 +2,4 @@ export type { EntryPoint } from 'https://deno.land/x/dnt@0.38.1/mod.ts';
 export { build, emptyDir } from 'https://deno.land/x/dnt@0.38.1/mod.ts';
 export { join, resolve } from 'https://deno.land/std@0.203.0/path/mod.ts';
 export { z } from 'https://deno.land/x/zod@v3.22.4/mod.ts';
+export { assertEquals, assertThrows, describe, it } from '../../dev_deps.ts';